project/requirements.txt
flask==2.2.*
joblib==1.2.*
sklearn==1.2.*
numpy>=1.23.*
pandas>=1.5.*
python-docx>=0.8.*
PyMuPDF>=1.20.*